+++ packages/wrapper/src/vscode/index.ts
@@ -1,0 +1,7 @@
+/* --------------------------------------------------------------------------------------------
+ * Copyright (c) 2024 TypeFox and others.
+ * Licensed under the MIT License. See LICENSE in the package root for license information.
+ * ------------------------------------------------------------------------------------------ */
+
+export * from './services.js';
+export * from './viewsService.js';